Initial configuration for Xenmen's sites
As of this point, the site will only be useful to me, and a pain for
anyone else to use since you'll have to be careful when git pulling.
However, this will be fixed next commit. In between now and then, I've
got lots more work to do on the site proper. Busy busy busy...